# exact `BigFloat` to IEEE FP conversion in pure Julia