@oscardssmith 2023-07-28
Thank you so much for writing this! I have a few minor style nits, but overall it looks really good.