@nsajko 2023-07-28
Note to myself: more tests would be good. An idea that comes to mind is to compare against `mpfr_get_flt` for all `Float32` values, varying through all rounding modes (both `RoundingMode` and `MPFR.MPFRRoundingMode`) and varying the `BigFloat` precision.