@nsajko 2023-07-28
Added more tests.