@nsajko 2023-07-29
Tweaked the new test in `test/rounding.jl`.